// src/app/pages/solicitudeprovider_products/components/solicitudeprovider_products-table/solicitudeprovider_products.interface.ts
export interface Solicitudeprovider_productsInterface {
   idsolicitudeprovider_product?: number;
   solicitudeprovider_idsolicitudeprovider?: number;
   product_idproduct?: number;
   quantity?: number;
   is_deleted?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}